Feat: Selección de institución y opción de salir
- Se añade la funcionalidad para seleccionar una institución.
  - Se crea `SeleccionarInstitucionFragment` y su respectivo ViewModel (`SeleccionarInstitucionViewModel`).
  - Se añade el layout `fragment_seleccionar_institucion.xml`.
  - Se añade el layout `item_institucion.xml` y se actualiza su adaptador `InstitucionAdapter` para permitir la selección.
  - Se integra la nueva pantalla en el grafo de navegación y en el menú lateral de `MainActivity`.
  - Se añade el icono `ic_menu_instituciones.xml`.
- Se añade la opción "Salir" de la aplicación en el menú lateral de `MainActivity`.
  - Se implementa un diálogo de confirmación antes de cerrar la aplicación.
  - Se añade el icono `ic_menu_salir.xml`.
- En `PreCargarActivity` y `PreCargarViewModel`:
  - Se cambia la visibilidad de `ProgressBar` y `Button` de `GONE` a `INVISIBLE` para mantener el espacio en el layout.
  - Se elimina un comentario innecesario en `PreCargarViewModel`.
- Se actualizan las cadenas de texto en `strings.xml`.
- Se actualiza el mapeo de VCS en `.idea/vcs.xml`.

+++ app/src/main/java/com/nutrizulia/presentation/viewmodel/PreCargarViewModel.kt
@@ -25,7 +25,6 @@ class PreCargarViewModel @Inject constructor(
     private val syncUsuarioIntituciones: syncUsuarioIntituciones
 ) : ViewModel() {
 
-    // --- LiveData existentes ---
     private val _mensaje = MutableLiveData<String>()
     val mensaje: LiveData<String> get() = _mensaje
 
@@ -68,7 +67,7 @@ class PreCargarViewModel @Inject constructor(
                 _mensaje.postValue("Sincronizando perfil de usuario...")
                 syncUsuarioIntituciones(usuarioId)
 
-                val perfiles = getPerfilesInstitucionales(usuarioId.toInt())
+                val perfiles = getPerfilesInstitucionales(usuarioId)
                 _profiles.postValue(perfiles)
 
                 if (perfiles.isEmpty()) {
